refactor: move directory locales
