+++ src/shared/config/locales/en.ts
