+++ src/shared/config/locales/index.ts
