+++ src/shared/config/locales/ru.ts
