+++ src/shared/utils/hooks/useTranslation.ts
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router'
-import { en, ru } from 'src/shared/config/i18n/locales'
+import { en, ru } from 'src/shared/config/locales'
 
 export const useTranslation = () => {
   const router = useRouter()
